+++ resources/js/lists/components/ListsControls.js
@@ -1,4 +1,4 @@
-import ListItemPicker from "../ListItemPicker";
+import ListItemPicker from "../controllers/ListItemPicker";
 
 class ListsControls {
