+++ resources/js/lists/index.js
@@ -1,0 +1,9 @@
+import ListItemPicker from "./controllers/ListItemPicker";
+
+import '../../css/style.css';
+
+// Create the app
+const itemPicker = new ListItemPicker();
+
+// Start the list walk through
+itemPicker.render();
