+++ resources/js/lists/models/list-item.js
