+++ resources/js/lists/models/list.js
